Startup values and boundary check
Replaced the random 1500 as the starting limits with my remotes values
shrunk by 10 or 15 at each end, on the fly calibration will take care of
expanding but now it's not critical to calibrate. also found a bug
causing all channels to have a 0 for the lower bound, which results in
an unusable mapping situation. Added a check that the value must be over
200 for it to have any bearing whatsoever.